orchestrator/requirements.txt
openai
weave
python-dotenv
redis
langchain-openai
langchain-core
langgraph
langchain-mcp-adapters
redis-mcp-server
